1->1.

+++ GAIN/DataModel.py
@@ -1,5 +1,7 @@
+#%%
 import tensorflow as tf
 import numpy as np
+
 
 #%% Helpers
 def createTestMaskNHints(data):
@@ -46,7 +48,7 @@ def createMask(data,maskRatio):
         0,1 matrix of the same shape as data
     '''
     print(tf.shape(data))
-    return tf.dtypes.cast((tf.random.uniform(tf.shape(data),minval=0,maxval=1)>(1-maskRatio)),dtype=tf.float32)
+    return tf.dtypes.cast((tf.random.uniform(tf.shape(data),minval=0,maxval=1)>(1.-maskRatio)),dtype=tf.float32)
 
 def create_random_mask(data,fixed_mask,maskRatio):
     '''
@@ -58,8 +60,8 @@ def create_random_mask(data,fixed_mask,maskRatio):
         0,1 matrix of the same shape as data
     '''
     [sample_size,dim]=tf.shape(data)
-    full_random_mask=tf.dtypes.cast((tf.random.uniform(tf.shape(data),minval=0,maxval=1)>(1-maskRatio)),dtype=tf.float32)
-    return tf.tile([fixed_mask],[sample_size,1])+full_random_mask*(1-fixed_mask)
+    full_random_mask=tf.dtypes.cast((tf.random.uniform(tf.shape(data),minval=0,maxval=1)>(1.-maskRatio)),dtype=tf.float32)
+    return tf.tile([fixed_mask],[sample_size,1])+full_random_mask*(1.-fixed_mask)
     
    
 
@@ -73,8 +75,8 @@ def createHint(mask,hintRate):
         hint_mask: mask for creating hints with 1,0 values. Same size as X.
         hints: hints matrix with 1,0.5,0 values. Same size as X.
     '''
-    hint_mask=createMask(mask,1-hintRate)
-    hints=hint_mask*mask+(1-hint_mask)*0.5
+    hint_mask=createMask(mask,1.-hintRate)
+    hints=hint_mask*mask+(1.-hint_mask)*0.5
     return hint_mask,hints
     
 def createHints(mask,hintRate):
@@ -123,18 +125,18 @@ class DataModel():
         if fix_mask is not None:
             self.fix_masks=tf.tile([fix_mask],[self.sample_size,1])
         else:
-            self.fix_masks=tf.zeros([self.sample_size,self.Dim])
+            self.fix_masks=tf.zeros([self.sample_size,self.Dim],dtype=tf.float32)
             
-        full_random_masks=tf.dtypes.cast((tf.random.uniform([self.sample_size,self.Dim],minval=0,maxval=1)>(1-p_miss)),dtype=tf.float32)
-        full_random_hints_masks=tf.dtypes.cast((tf.random.uniform([self.sample_size,self.Dim],minval=0,maxval=1)>(1-p_hints)),dtype=tf.float32)
+        full_random_masks=tf.dtypes.cast((tf.random.uniform([self.sample_size,self.Dim],minval=0,maxval=1)>(1.-p_miss)),dtype=tf.float32)
+        full_random_hints_masks=tf.dtypes.cast((tf.random.uniform([self.sample_size,self.Dim],minval=0,maxval=1)>(1.-p_hints)),dtype=tf.float32)
         
         train_size=int(self.sample_size*train_rate)
         train_batch=int(train_size*batch_ratio)
         self.train_batch_size=int(train_batch)
 
-        masks=(1-self.fix_masks)*full_random_masks+self.fix_masks
-        hint_masks=(1-self.fix_masks)*full_random_hints_masks+self.fix_masks
-        hints=(1-hint_masks)*0.5+hint_masks*masks
+        masks=(1.-self.fix_masks)*full_random_masks+self.fix_masks
+        hint_masks=(1.-self.fix_masks)*full_random_hints_masks+self.fix_masks
+        hints=(1.-hint_masks)*0.5+hint_masks*masks
         
         fix_mask_set=tf.data.Dataset.from_tensor_slices(self.fix_masks)
         masks_set=tf.data.Dataset.from_tensor_slices(masks)
@@ -151,7 +153,7 @@ class DataModel():
         if train_rate!=1:
             test_size=self.sample_size-train_size
             test_batch=int(test_size*batch_ratio)
-            dataset_test=dataset_with_masks_and_hints.skip(train_size).batch(test_batch,drop_remainder=True).repeat(int(repeat*train_rate/(1-train_rate)))
+            dataset_test=dataset_with_masks_and_hints.skip(train_size).batch(test_batch,drop_remainder=True).repeat(int(repeat*train_rate/(1.-train_rate)))
             return dataset_train,dataset_test
         
         
@@ -171,7 +173,7 @@ class DataModel():
             Missing filled data.
         """
         x=self.normaliser(data)
-        return self.denormaliser(generator(x,mask)*(1-mask)+mask*x)
+        return self.denormaliser(generator(x,mask)*(1.-mask)+mask*x)
 
     def discriminate(self,discriminator,data,mask):
         """
@@ -187,6 +189,6 @@ class DataModel():
             the data is genuine but not generated.
             genuine -> 1, generated -> 0.
         """
-        hints=mask+(1-mask)*0.5
+        hints=mask+(1.-mask)*0.5
         x_hat=self.normaliser(data)
         return discriminator(x_hat,hints)
